Feat: 피드리스트

+++ backend/raonzena/src/main/java/com/ssafy/raonzena/api/response/UserRes.java
@@ -9,7 +9,7 @@ import org.springframework.beans.BeanUtils;
 @Getter
 @NoArgsConstructor
 public class UserRes {
-    private int userNo;
+    private long userNo;
     private String userId;
     private String userName;
     private int exp;
